server separeted from client
