+++ client/public/client.html
@@ -6,7 +6,6 @@
   <body>
 	<p id="socketId"></p>
 	<pre id="logs"></pre>
-	<script src="/socket.io/socket.io.js"></script>
-	<script type="text/javascript" src="client.js"></script>
+	<script type="module" src="client.js"></script>
   </body>
 </html>
