+++ server/.eslintrc.json
@@ -1,0 +1,59 @@
+{
+	"globals": {
+		"document": "readonly",
+		"io": "readonly"
+	},
+	"env": {
+	  "node": true,
+	  "commonjs": true,
+	  "es2021": true,
+	  "jest": true
+	},
+	"extends": "eslint:recommended",
+	"parserOptions": {
+	  "ecmaVersion": 12
+	},
+	"rules": {
+	  //* Avoid Bugs
+	  "no-undef": "error",
+	  "semi": "error",
+	  "semi-spacing": "error",
+	  //* Best Practices
+	  "eqeqeq": "warn",
+	  "no-invalid-this": "error",
+	  "no-return-assign": "error",
+	  "no-unused-expressions": ["error", { "allowTernary": true }],
+	  "no-useless-return": "error",
+	  "no-constant-condition": "warn",
+	  "no-unused-vars": ["warn", { "argsIgnorePattern": "req|res|next|__" }],
+	  //* Enhance Readability
+	  "indent": ["error", "tab"],
+	  "no-mixed-spaces-and-tabs": "warn",
+	  "space-before-blocks": "error",
+	  "space-in-parens": "error",
+	  "space-infix-ops": "error",
+	  "space-unary-ops": "error",
+	  "quotes": ["error", "single"],
+	  //
+	  "max-len": ["error", { "code": 400 }],
+	  "max-lines": ["error", { "max": 500 }],
+	  "keyword-spacing": "error",
+	  "multiline-ternary": ["error", "never"],
+	  "no-mixed-operators": "error",
+	  //
+	  "no-multiple-empty-lines": ["error", { "max": 2, "maxEOF": 1 }],
+	  "no-whitespace-before-property": "error",
+	  "nonblock-statement-body-position": "error",
+	  "object-property-newline": [
+		"error",
+		{ "allowAllPropertiesOnSameLine": true }
+	  ],
+	  //* ES6
+	  "arrow-spacing": "error",
+	  "no-confusing-arrow": "error",
+	  "no-duplicate-imports": "error",
+	  "no-var": "error",
+	  "object-shorthand": "off",
+	  "prefer-const": "error"
+	}
+  }
