+++ server/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "server",
+  "version": "1.0.0",
+  "description": "server side of the logger app",
+  "main": "server.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node server.js",
+	"lint": "eslint ./"
+  },
+  "author": "Leandro",
+  "license": "ISC",
+  "dependencies": {
+    "dotenv": "^16.3.1",
+    "eslint": "^8.48.0",
+    "express": "^4.18.2",
+    "nodemon": "^3.0.1",
+    "socket.io": "^4.7.2",
+    "socket.io-client": "^4.7.2"
+  }
+}
